[MOD/#24] 파일 분리 및 네이밍 변경
