+++ core/designsystem/src/main/java/com/hilingual/core/designsystem/component/dialog/DialogButton.kt
@@ -1,0 +1,47 @@
+package com.hilingual.core.designsystem.component.dialog
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.unit.dp
+import com.hilingual.core.common.extension.noRippleClickable
+import com.hilingual.core.designsystem.theme.HilingualTheme
+
+@Composable
+fun DialogButton(
+    text: String,
+    modifier: Modifier = Modifier,
+    onClick: () -> Unit,
+    isFilled: Boolean = true,
+) {
+    Row(
+        modifier = modifier
+            .clip(RoundedCornerShape(8.dp))
+            .background(
+                color = if (isFilled)
+                    HilingualTheme.colors.hilingualOrange
+                else
+                    HilingualTheme.colors.gray100
+            )
+            .padding(vertical = 14.dp)
+            .noRippleClickable(onClick = onClick),
+        horizontalArrangement = Arrangement.Center,
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Text(
+            text = text,
+            style = HilingualTheme.typography.bodySB16,
+            color = if (isFilled)
+                HilingualTheme.colors.white
+            else
+                HilingualTheme.colors.gray400
+        )
+    }
+}
